Fix issues reported by lein eastwood
The linter found another misplaced docstring in inclusive-seq, and what
appears to have been an actual bug in heat-hsl:

== Linting com.evocomputing.colors.palettes.core ==
{:linter :misplaced-docstrings,
 :msg "Possibly misplaced docstring, inclusive-seq",
  :file "com/evocomputing/colors/palettes/core.clj",
   :line 16,
    :column 7}

{:linter :unused-ret-vals,
 :msg
  "Constant value is discarded inside inclusive-seq: \"Return n evenly
  spaced points along the range start - end (inclusive)\"",
   :file "com/evocomputing/colors/palettes/core.clj",
    :line 16,
     :column 1}

{:linter :suspicious-expression,
 :msg
  "* called with 1 args.  (* x) always returns x.  Perhaps there are
  misplaced parentheses?",
   :file "com/evocomputing/colors/palettes/core.clj",
    :line 142,
     :column 45}

Also fix an inconsistent reference to the old location of the string
library which snuck in as I jumped back and forth between versions.
Clojure 1.7 is in release candidate stage, so hopefully we can stop
doing that soon and I can release all my changes!