No longer crash when printing some HSL colors.
Formerly, if you accidentally passed integers for the saturation or
lightness values when creating a color using an HSL map, the color would
create fine, but then when anything tried to print it (for example,
looking at a structure containing it in the REPL), an exception would be
thrown, because the format string for printing the color was expecting a
float.

This change adds a test which reveals that exception, and then adds a
fix which allows you to pass integers for HSL components, and explictly
promotes them to floats during the construction process, which is more
consistent with other aspects of the Clojure ecosystem, and certainly
less astonishing than having colors which seem to work until you try to
print them.