Fix compiler warnings with current versions of Clojure
The `seqable?` function is now part of clojure.core, so including this
older version from clojure.contrib is causing compile-time warnings.
This also eliminates the need to depend on clojure.contrib at all.

I also removed `lein-ancient` from the project plugins. The version
referenced crashes with current Java and Clojure distributions, and
this tool is something that better fits in an individual's
`~/.lein/profiles.clj` rather than in the project definition.