CLJS port
Add a shadow-cljs test config for browser unit tests
Use reader conditionals to exclude some Clojure specific code (x11, awt).
In namespace forms remove (:use) which is unsupported on clojurescript.
Move macros to a separate namespace (https://code.thheller.com/blog/shadow-cljs/2019/10/12/clojurescript-macros.html).
Create a utils and a math namespace to wrap some commonly used functions in a
platform independant manner.
Change deprecated defstruct for color to a defrecord. This has the added benefit
of working on both platforms.
Fix an issue on Clojurescript with rgba-int-to-components due to a call to
bit-shift-right which behaves differently from Clojure. I think the original
code meant to use unsigned-bit-shift-right instead so use that. It still works
on Clojure and fixes the Clojurescript issue. Add a corresponding unit test.
All tests passing both on Clojure and Clojurescript.